Add test environment without any extras installed
