+++ .github/workflows/ci.yml
@@ -34,6 +34,8 @@ jobs:
         run: python3 -m pip install tox tox-uv
       - name: run tox
         run: tox -e 'py${{ matrix.python_version }}'
+      - name: run tox (base-only tests)
+        run: tox -e 'py${{ matrix.python_version }}-base-only'
       - name: submit code coverage
         uses: codecov/codecov-action@v4
         env:
