+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py{39,310,311,312}
+envlist = py{39,310,311,312}{,-base-only}
 
 [testenv]
 commands=
@@ -18,6 +18,11 @@ setenv=
     # See https://github.com/pytest-dev/pytest/issues/2042
     PY_IGNORE_IMPORTMISMATCH=1
 
+[testenv:py{39,310,311,312}-base-only]
+# deliberately left blank! this environment is meant
+# for testing without any extras supplied
+extras=
+
 [testenv:numba_coverage]
 commands=
     pytest --durations=5 --cov=libertem_blobfinder --cov-report=term --cov-report=html --cov-report=xml --cov-config=setup.cfg -m with_numba {posargs:tests/}
@@ -34,7 +39,7 @@ skip_install=True
 
 [testenv:docs]
 changedir={toxinidir}
-basepython=python3.7
+basepython=python3.9
 whitelist_externals=
     make
 setenv=
